requirements.txt.txt
streamlit==1.24.0
python-binance==1.0.16
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
joblib==1.3.2
matplotlib==3.7.2
requests==2.31.0